agent-local/requirements.txt
livekit-agents[openai,deepgram,silero,groq]==1.4.1
python-dotenv==1.0.1
aiohttp==3.12.15
